requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
idna==3.7
mutagen==1.47.0
pillow==10.3.0
psycopg2==2.9.9
pycryptodomex==3.20.0
requests==2.32.2
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
yt-dlp==2024.5.26
